Make DynamicStockModel's inherit from Stock
