+++ sodym/__init__.py
@@ -4,4 +4,4 @@ from sodym.mfa_definition import (
 from sodym.mfa_system import MFASystem
 from sodym.dimensions import DimensionSet, Dimension
 from sodym.named_dim_arrays import NamedDimArray, Parameter, Process, StockArray, Flow
-from sodym.stocks import StockWithDSM, Stock
+from sodym.stocks import Stock, DynamicStockModel
